Added config for travis-ci auto-builds
This adds the .travis.yml file that tells Travis what and how to build.
See http://docs.travis-ci.com/user/build-configuration/ for docs on how
the directives in the file configure the build.

Added BUILD_KERNEL to Travis build spec

The Ubuntu 12.4 LTS images Travis uses run a kernel too old for the igb
library or module to build, so we need to build against an installed but
newer-than-running kernel. The BUILD_KERNEL option tells the module build
to use the alternate kernel source.

Add install of ethtool.h to system includes to Travis config

Again, the Ubuntu 12.4 LTS image has an old kernel's userspace API headers
installed in /usr/include/linux, and this copies one of the headers from
the newer kernel we're building against into the userspace API area.

Added libsndfile1-dev to Travis config for simple-listener build