Small changes to test and CI infrastructure
Travis is now configured such that ctest will be invoked with the
--output-on-failure option, which will dump the test output from the
failing test.

The cmake configuration is also configured to call the test program with
the -v option, which causes it to include the test names in the output as
they run.

The end result should make it much easier to debug problems that occur
during CI builds.