Handle link up/down events
Created a link_thread to monitor the NIC status;
Created two extra events on IEEEPort::processEvents to
handle the link up/down state transitions;
On link up, the HWTimestamper init is called. On link down, it is
just being logged the state transition.

This commit is based on the automotive-profile branch and fixes Issue #254.