Starting with Wireshark 2.0.5 the signaling message in gPTP is reported as Malformed due to an incorrect value for the LengthField.
As defined in 802.1AS 2011 section 10.5.4.3.3 LengthField the length value shall be 12.
Currently gPTP is sending an incorrect value of 28.

Additionally for the signaling message TLV: Organization Subtype is 0x1 in the packet but it should be 0x2
LinkDelayInterval is 0x0 but for the Automotive Profile should be 127
Flags field is 0x0 in the packet but should be 0x3