gPTP: Remove timestamper_init call on link up
The Linux hardware timestamper init is only designed to be called
once upon creation. Calling it on every link up event results in
resource leaks. Therefore the call has been removed.

A substitute, timestamper_reset, has been added in its place. It
currently only calls the init_phy_delay function, since there are
no other exposed methods in the HWTimestamper interface that were
clearly appropriate.

We may also want to add a call to reset the frequency adjustment,
but that is not currently available in the HWTimestamper interface.