After yielding, automatically reserve another block
The application will automatically attempt to reserve another block of
the same size after yielding a block it was defending.  The ID will be
the same so that the reserving application can track the new
reservation.
The status handling code has been updated to not return the status for
an ID that has not yet been announced.
Includes updated unit tests.