Descriptor handing fixes and optimizations
Added optimizations when looking for descriptors that are not part of a
CONFIGURATION (non-top-level).
Verify the descriptor index value before returning an ENTITY or
CONFIGURATION descriptor.
Added openavbAemGetDescriptorIndex() function.
Changes to not add the same STRINGS descriptor more than once, and
report an error when attempted.
Debugging output shows descriptor types as hexadecimal, to match 1722-1
format.