INI file parsing updates
Ignore Class A and Class B for listeners, as it is not specified.
Record the intf_nv_audio_rate, intf_nv_audio_bit_depth, and
intf_nv_channels for reference.
Use the intf_nv_audio_rate value to specify the current_sampling_rate,
if not specified.
Abort parsing the .INI file if an error is found.
Added more error messages if issues found parsing the .INI file.