Client uses .INI file name to connect to AVDECC server
The client is now using a "friendly_name" that defaults to the name of
the .INI file to associate with the server.  This will make it easier
once hard-wired information (Talker MAC, Stream ID, etc.) is removed or
not specified in the .INI file.
Added changes to openavbAvdeccMsgClntNotifyCurrentState() to abort
failed searches earlier.