Put igb-dependent code behind a build feature in avtp_pipeline
This is a bit of refactoring so that the code that depends on libigb is
separated from the rest of the Linux platform code. Different platform toolchain
files for cmake are now selectable; the following command will build based on
the "generic.cmake" toolchain file, which does not include igb support.

$ PLATFORM_TOOLCHAIN=generic make avtp_pipeline