Changed the ACMP_SM_UNLOCK timing for incoming packets
Changed the ACMP_SM_UNLOCK timing to ensure the incoming packet being
processed is not changed during state machine processing triggered by
that packet.  This change may not be necessary, but is easier to prove
that it will work as desired.  It also matches the style used in
openavb_aecp_sm_entity_model_entity.c.