Added user support for setting the ADP valid_time
The ADP valid_time defaults to 62 seconds (as specified in 1722.1-2013
clause 6.2.1.6), which means that devices remain valid on AVDECC
controllers on the network for over a minute after the device crashes.
This change allows the user to specify the valid_time in avdecc.ini, and
has the default version of that file specify a valid_time of 20 seconds.