Added support for the MAAP daemon
Use the MAAP daemon to get the multicast destination address, if
available.  If an address is needed, and MAAP is not available, allocate
one from the MAAP locally administered Pool.
Have the Talker/Listener keep track of the AVDECC Msg handle, and use it
to notify AVDECC Msg when the multicast destination address changes
(i.e. is assigned by MAAP).