AVDECC waits for stream ID before sending GET_TX_CONNECTION_RESPONSE
When using MAAP, the destinatation address is not allocated until the
endpoint is notified by the Talker that the stream should be started.
AVDECC now waits until it is notified of the MAAP-allocated destination
address before sending the GET_TX_CONNECTION_RESPONSE command (in
response to a GET_TX_CONNECTION_COMMAND command).