Added AVTP Pipeline support for the Shaper daemon
Added support for the AVTP Pipeline to interact with the Shaper daemon.
The support defaults to being disabled, and can be enabled in the endpoint.ini
file.
If available and support is enabled, the Shaper daemon is configured when a
Talker stream starts, and unconfigured when the stream ends.
Note that the AVTP Pipeline version number was incremented, as the interface
between the Talker/Listener and the Endpoint has been updated.