Added ACMP saved state support
When the AVTP Pipeline with AVDECC support is told by the controller to run as
a Listener, the connection information is saved to avdecc_save.ini.  The saved
state information is deleted when the connection is closed by the AVDECC
controller.  This will allow for future fast connect support that can used the
saved state when AVDECC is initialized.