Initial Fast Connect Support
Saves additional information (flags and talker_unique_id) needed to support
fast connect.
SAVED_STATE flag set during the initial connection.
Fast Connect is attempted if there is saved state for the Listener when it
connects to the AVDECC app.
Second timeout to Talker on Fast Connect (as defined in IEEE 1722.1-2013
Clause 8.2.2.1.1) still needs to be implemented.