Fast connect retried if Talker is discovered
If the initial fast connect attempt has timed out, watch for ENTITY_AVAILABLE
discovery PDUs to indicate that the Talker is now available.  Reattempt the
fast connect to the Talker once an appropriate PDU is detected.
Keeps track of the fast connect status in the
openavb_aem_descriptor_stream_io_t descriptor, for easy access.
Renamed openavbAvdeccSaveStateInfo() to openavbAvdeccGetSaveStateInfo() to
make the purpose more clear.