Received AECP messages are queued
Rather than trying to process each AECP message as the AEM command is
received, the messages are queued, and processed in the order they are
received.  This simplifies the multithreading synchronization logic, and
doesn't rely on the network support to queue a large number of frames is
several messages are sent in a row.