Fix if Talker disappears and reappears
If the Listener determines that the Talker it was connected to has stopped, it
waits 5 seconds before attempting a fast connect to the Talker.  This is to
deal with cases where the Talker would still advertise itself and/or respond
to CONNECT_TX_COMMANDs while in the process of shutting down.
Stop the fast connect retry if the saved state information is deleted.
Also, fix to avoid divide-by-0 error seen occasionally when registering the
stream at a bad time.