STATUS_NOT_SUPPORTED returned for unsupported commands
For the commands SET_STREAM_FORMAT, SET_SAMPLING_RATE, and SET_CLOCK_SOURCE,
instead of just updating the descriptors, added a TODO noting that the
supplied value needs to be verified and the Talker/Listener needs to be
notified of the change.  For now, STATUS_NOT_SUPPORTED is returned in those
situations.