Functions changed to return the current state, rather than IsStreaming
The openavbAVDECCListenerIsStreaming() and openavbAVDECCTalkerIsStreaming()
functions have been replaced with openavbAVDECCGetRequestedState() and
openavbAVDECCGetStreamingState(), which allows for more flexibility.
Changed the value of OPENAVB_AVDECC_MSG_STOPPED_UNEXPECTEDLY so that
OPENAVB_AVDECC_MSG_RUNNING or greater values will indicate that the client is
active.  The version number was updated to reflect this change.