Added SET_STREAM_INFO implementation
When SET_STREAM_INFO is used, settings are sent to the client.
Client uses those settings, or restores from the backup of the original
settings, as appropriate.
If the destination address is specified, then MAAP allocation will not be
used.
GET_STREAM_INFO updated to reflect the settings supplied for the Talker, and
the settings currently in use for the Listener.