AAF Integer Conversion Support Added
When using AAF in 16-bit, 24-bit, or 32-bit integer mode, the incoming audio
can be 16-bit, 24-bit, or 32-bit integer data.  The incoming audio is
converted to the desired internal format.
Note that the conversion only works correctly if the incoming audio is big-
endian, which IEEE 1722-2016 Clause 7.3.5 lists as a requirement for AAF.
INI files updated to match IEEE 1722-2016 specification.