Daemon scripts cleanup
Added shaper_daemon to run_daemons.sh and stop_daemons.sh.
Fixed the path and parameters for maap_daemon.
IGB is now initialized separately, as the daemons can be run without it.
Don't assume the name of the NIC, and don't include "sudo" in the script, to
be consistent with the other run_*.sh scripts.