Added EXTENDED_SEQUENCE_NUMBERS option
Added support for extended sequence numbers, which is a non-standard way to
increase the number of bits (from 8 to 12) for sequence numbers in AVTP
frames.  This is useful for Wi-Fi with retries, which can drop large numbers
of sequential frames.
To enable this feature, uncomment the EXTENDED_SEQUENCE_NUMBERS define.  The
feature is currently disabled.